fix: Fix client side
 ## Problem
I added a fix in #36 (https://github.com/redwoodjs/reloaded/pull/36/commits/90c3734398b72fd3b04de0c4b360043918c44596) that caused us to always resolve react to a custom
build of react. This custom build caters for the
fact that RSC and SSR are sharing a runtime.

However, I accidentally made the client side bundling also use this custom build of react, rather than normal react.

 ## Solution

I've added a plugin that allows us to alias react to the custom build of react only for the worker side, but leave it alone on the client side.

 ## Note
While fixing this, I discovered that the current way of having RSC and SSR share a single react build will not work for all things. For example,
if there is a `use()` call in SSR-rendered code, we will end up with an issue because RSC is expecting its `use()`` and not the SSR `use()`.

I'll work on fixing this in a future PR.