fix: Allow for cloudflare worker handlers other than fetch (#61)
## Context
In order to get vite HMR and miniflare to work together, we need to have a wrapper entry point module, which will import and run the relevant code dynamically, leveraging vite dev server for loading the modules.

The way we were defining this wrapper was with a durable object. We exposed methods on the durable object for initialization (we need to create a vite "module runner" inside miniflare), as well as sending messages from vite into miniflare. This way, we can get at the "stub" of the durable object instance outside of miniflare, to call those methods via RPC. In other words, defining methods on a durable object and then calling them from outside of miniflare is one way to hook vite dev server up to code running inside of miniflare. This approach is what the vite team use in [their workerd example](https://github.com/hi-ogawa/vite-environment-examples/blob/81f30c9f9618878fd823a28d5970d6432d0d7002/packages/workerd/src/worker.ts#L15).

## Problem
The problem with this approach, is that DurableObjects don't expose the same handlers as workers do. They both can provide a `fetch()` handler, but for e.g. workers have a `queue()` handler and DurableObjects don't.

As a result, we weren't able to use things like `queue()` in dev.

## Solution
[Cloudflare's vite plugin](https://github.com/flarelabs-net/vite-plugin-cloudflare/blob/2876d044a1d815326b076175a7d1f3444ee47c59/packages/vite-plugin-cloudflare/src/runner-worker/index.ts) contrasts this approach by instead defining an actual worker when wrapping workers, and wrapping around DO's for DO's  (sensible right). As such, they don't have the same issue.

So in this PR, we solve the problem the same way as CF do basically.

Since we no longer have a durable object as the main wrapper, we now need a new way of calling methods that are in the miniflare sandbox from outside of miniflare. The way this PR solves this is by issuing these as internal requests to the wrapper worker, which instead of proxying those requests to the wrapped worker, do the internal work we're wanting it to do (e.g. instantiate the vite module runner)