fix: Avoid env exposure for client-based pages (#127)
## Problem
When rendering page components (the "root" component for a page), we provide it with props that includes `env`. This is fine when the page component is a worker-rendered RSC component, but not fine when the page component is a `use client` component - in the latter case, we are exposing `env` to the browser.

## Solution
Detect if page component is a client component. If it is, pass only `params` and `ctx` as props.

**note** We still need to be careful to not provide anything in ctx that we do not want shown to the browser. I'm not sure if it would be better for us to not pass `ctx` down either. But then we maybe want to prevent page components from being `use client` components in the first place. Or maybe we can make it clear in docs that ctx passed to client side?