fix: Avoid unnecessary passing of Context type param (#137)
## Changes
### 1. Removed unnecessary passing of `Context` param
The session and passkey-auth starters had a bunch of unnecessary passing of `Context` param to route API methods, for example:

```ts
export default defineApp<Context>([
  async ({ env, ctx, request }) => {
    setupSessionStore(env);
    ctx.session = await sessions.load(request);
  },
  // >>> Context here not needing to be passed
  layout<Context>(Document, [
    index([
        Home,
    ]),
    // >>> also here
    prefix<Context>("/user", authRoutes),
  ])
])
```

At some point we needed these, but we don't seem to need it anymore (without losing type information). TBH I'm not sure what changed, but glad we don't need to do this anymore regardless.

### 2. Remove unnecessary explicit type annotations for routes
#### Problem
We were needing add explicit type annotations for arrays of routes, in order for ts to accept them being provided as part of `defineApp`, e.g.

```ts
const authRoutes: RouteDefinition<Context> = [...]
```

This is because the default for the `Context` type param is `undefined` when it cannot not be inferred (which happens for the array-of-routes case here).

#### Solution
To solve this, we default `Context` type param to `any` in the routing API's methods. This way:
* when it can be inferred, the inferred type is used
* if it cannot be inferred, we also might not need to care (e.g. if routes are referencing `ctx`), so rather default to `any` to avoid needing to add explicit type annotations
* if it cannot be inferred, in the cases where do need to care, we still can pass the type explicitly