fix: Fix useClient transform for non-inlined exports (#138)
## Problem
Part of our 'use client' transformations for worker-side code involves renaming components to include an `SSR` suffix. After #126, we do this transformation by locating the line+column where the `export` is. This is fine for inlined exports (`export function Foo() {...}`) but breaks for grouped exports (`function Foo() {}; export { Foo }`).

## Solution
Find where the function is defined and replace it there.