feat: Symlink .env to .dev.vars automatically (#153)
## Problem
At the moment, one needs to put env vars in `.dev.vars` for `@cloudflare/vite-plugin` to pick them up.

Thing is, people expect `.env` to be where things are put.
* There's been two issues already for us getting tripped up on this: #97, #142
* Its become a well-established convention, to be able to put things in `.env`

## Solution
We symlink from `.env` (if it exists) to `.dev.vars` (if it does not already exist).

We do not handle removing of the symlink, since it is tricky to ensure this always happens.

## ⚠️ Worth noting
These vars **won't** be used for deploys. In this case, one needs to use `wrangler secrets put ENV_VAR` ([wrangler docs on secrets](https://developers.cloudflare.com/workers/configuration/secrets/))