fix: Default type for route params shouldn't restrict to Record<string, string> (#188)
## Problem
We were setting the default type for route params to `Record<string, string>`, and then not providing type params across the relevant types for `route()`. This means that instead of ts being able to infer the type for params from the component passed to `route()`, ts would compare the type for params for the component to `Record<string, string>`, (e.g. for `const Component = (props: RouteContext<Context, { foo: string }>) => ...` to would compare `{ foo: string } `to `Record<string, string>`).

## Solution
* Provide type params for route params across relevant types for `route()` - to make inference possible
* For types that contain multiple routes (.e.g `layout`), use `any` as a default for the route params - since it might be more technically correct if we can get ts to infer the type for params to be a union type at that level, but I'm not sure if there's a practical difference from user's perspective for those cases