chore: Rename RouteContext to RouteOptions (#189)
## Problem
Currently we call the properties we pass to routes `ctx: RouteContext`, but we also have a `ctx` key inside of it (for the user's own mutable context object). This means we have a `ctx.ctx` - this looks weird, and it is easy to mistake the one `ctx` for the other this way.

## Solution
Rename `ctx: RouteContext` to `opts: RouteOptions`:

```ts
const SomeComponent = (opts: RouteOptions) => {
  const { env, request, params } = opts
  ...
}
```


## ⚠️ Breaking Change
Any references to `RouteContext` in your app need to be renamed to RouteOptions`