fix: Run worker scripts sequentially (#202)
## Context
We have a way of running scripts with code run inside of workers - this what `pnpm seed` uses (so that it can use d1 the same way an application would for populating seed data).

Our "worker script" runner relies on CF's vite plugin, the same as what we use for `pnpm dev`. This plugin relies on Miniflare, which starts up something called an inspector on port 9229. It is not currently possible to configure this port through the vite plugin, though with a small PR it could be made simple enough to do, as Miniflare itself does appear to support the inspector port number as an option.

We also have a postinstall hook in our starters, that will migrate the database and run the seed script.

In our monorepo, pnpm runs each starter's postinstall concurrently, and doesn't offer configuration for changing this behaviour. They offer a cli flag, but it isn't practical for us to need to provide this flag everywhere we need to `pnpm install`

## Problem
The concurrent pnpm installs cause concurrent worker script runs, which cause multiple inspector port listen attempts to happen on 9229 concurrently. Implications
1. Our own pnpm install (e.g. on CI) are broken until we fix this
2. Users can't run worker scripts concurrently

## Solution
Queue worker script runs to happen sequentially. We use a lockfile rather than waiting for 9229 to open up in order to mitigate a stampede of port listens when the port does eventually free up.