fix: `export default` cases for "use client" (#211)
## Context
For `"use client"`, the sdk works for inlined `export` (e.g. `export const function Component() {}` and grouped exports `export { Component }`, but does not work for `export default cases`. This PR gets it working with the export default cases.

## Implementation
Up until now, we've been using regexes as a way to do the necessary transformations for `use client`. This was fine in the beginning, but in order to cater for all the different kinds of export cases (grouped exports, inlined exports, grouped default exports, inlined default exports, and for each of those for arrows, functions, and for each of those sometimes anoynmous, sometimes not), it has grown in complexity to the extent that regexes are no longer a reliable way of extracting the necessary information and transforming the code, and its just getting out of hand.

We now use `ts-morph` (a thin wrapper around ts's compiler API) to parse and transform the ast for use client cases.