fix: Realtime updates given in action responses (#232)
In our realtime API, there are two cases where need to update the UI with a new RSC payload stream:
* 1. rsc update: a change issued by a different client (or by server) has  resulted in a new rsc payload
* 2. action result: a client called an action, so it gets a new rsc payload, including the UI update and action result

We were only setting a new payload stream in case 1 and not case 2.

Our note taking demo did not catch this, since the text was already being changed client side since it was a text box!