fix: Avoid worker import in router (#267)
## Problem
Now with #251, the sdk's router runtime code imports code in the module where `requestInfo` is defined, which in turn uses async_hooks (a node API, which CF also supports and polyfills). The problem comes in when client side code tries import `@redwoodjs/sdk/router` - when it needs to do to use `defineLinks()` (valid for that to be client side). As a result, the client side tries import async_hooks.

## Soution
Avoid imports requestInfo module in router