fix: Skip calling Page functions for client references (#345)
Since #342, we call page functions rather than letting react unroll the element tree from that level, in order to check whether the result is a response or an element.

Problem is, the Page function may be a `use client` component, in which case:
* it will never be able to return a `Response`
* The hooks that need to work in the context of ssr (e.g. `useState()`) are not set to use the correct internal react state used by ssr) if called as a function like this

To solve this, we avoid calling page functions if they are client component (since we won't be getting a `Response` from them anyways).