fix: Avoid info collection after transforms for use client export defaults (#352)
## Context
When transforming `use client` modules for worker environment
* we keep the original components code as the SSR form the components (and export under name `<Component>SSR`) - this is used for SSR rendering
* we transform the existing component exports to instead be registered client references (since the actual component code is only relevant to client environment)

## Problem
In #346, we fixed the transformations we were doing for grouped exports via making the code more robust, by collecting all info we need before doing any transformation. Except in attempting to make it more robust I introduced a new regression for the `export default case`, since we are now getting some info _after_ transformation for that case.

## Solution
Ensure we are _actually_ collecting everything we need before transformation, to also solve the regression for the `export default` case.