fix: More robust Page: () => Element | Response (#354)
## Context
We support pages returning either an element, or a response. The way we do this, is by simply calling the page component as a function, and inspecting whether the result is an element or a response

## Problem
The issue with simply calling the page component, is that this happens _outside of the react rendering context_. React does some set up to its internal state at the beginning of rendering. I think it is likely we will run into issues because of this. e.g. if there are RSC-valid hook usages (e.g. `use()`) or maybe even `Suspense`s.

## Solution
We want to avoid unrolling this first level in the component tree and leave this up to react.

We can do this by wrapping around the component, and calling it within the wrapper. If we are instantiating the wrapper as a react element and leaving react to unroll it, that contained page component call works since we are now in react's rendering context. When we call the component, we can then inspect its result. We then `throw` to bubble up the response (something we already support).

Basically:

```tsx
const PageWrapper = async () => {
  // calling Page() here works because it will happen within react rendering context
  const result = await Page(props);

  if (result instanceof Response) {
    // throwing allows us to bubble up the response and break out of render
    throw result;
  }

  return result;
}

renderToReadableStream(<PageWrapper />)
```