fix: Client entrypoint 404s (#400)
## Context
We have two vite environments: `client` (for client side) and `worker` (for code run in cloudflare worker).

The build step of `worker` relies on the build manifest vite generates for `client` - this asset includes mappings from the entry point source paths to their corresponding built destination paths.

It relies on the manifest so that we can replace references to client entry points in jsx (e.g. in `Document.tsx`) with their corresponding built assets.

## Problem
We read the client manifest from disk the first time, then hold onto it in memory and use the in memory value instead for subsequent attempts to read it. However, when reading the client manifest, we do not check whether the current context is the `worker` environment (not the `client` environment). This causes two issues:
* if this is the first build/release for the project on your machine, there won't be a manifest yet by the time we read => we'll default to an empty lookup for it (`{}`)
* if you had released before with different code, the new manifest hasn't been generated yet by the time we read => we'll use an outdated manifest
  * Seen here
     * #378 
     * #399

## Solution
* Change the plugin that relies on the manifest to only run for the `worker` environment (it should not be running for `client` anyways).
* We also now throw rather than default to `{}` so that we can catch issues if for some reason we are still needing the manifest before it exists