Rsc bundling implementation v2 (#488)
## Problem 1: RSC+SSR react runtimes

Since we started with RedwoodSDK, there's been one recurring challenge: getting React's RSC runtimes and SSR runtimes to coexist in the same environment - a cloudflare worker.

Our current solution involves making a custom react build, where the react internals required for both RSC and SSR are present for them to both work. While this solution works, it obviously is a hack, and there's no guarantee we're not going to hit any walls.

## Problem 2: `react-server` condition

This custom react build also allowed us to work around the fact that RSC react and SSR react are exposed using different export conditions:

However, this only solves things for react. Part of being a valid RSC implementation involves supporting packages that have exposed their own `react-server` export:
https://github.com/reactjs/rfcs/blob/main/text/0227-server-module-conventions.md#react-server-conditional-exports

It is likely that as RSC picks up adoption, more and more libraries will include this export.

## Problem 3: `"use client"` in `node_modules`

With RedwoodSDK in its current form, we are only able to support `"use client"` and `"use server"` directives given in the application code itself. Any packages with modules containing these directives are not currently supported in RedwoodSDK.

## Background Context
Vite supports separate environments, where different options can be applied to each environment for both dev and production. This gives us a natural opportunity to configure a separate SSR environment for the module resolution and transformation that should happen for the SSR work that happens on the worker.

We currently have two environments: `client` (for code run in browser) and `worker` (for code run in cloudflare worker), where `worker` environment uses `@cloudflare/vite-plugin` to get the code to run in a cloudflare worker and to configure it to be runnable there.

This is where things get tricky though - RSC requires us to configure two different environments differently (resolve react deps to different places, transform code differently depending on whether we're dealing with SSR renders or RSC). Yet this all has to go through that single `worker` environment, since ultimately evertyhing is bundled to run in the same cloudflare worker runtime environment. So for RSC we need two environments, but we somehow need to get them to eventually run in a single environment.

## Solution
### The SSR Bridge
This PR adds an `ssr` environment with bundler logic (module resolution, transformation) specific to SSR. We then introduce the concept of the _SSR bridge_: this is the entry point to all SSR code. When the `worker` environment loads the bridge module (`rwsdk/__ssr_bridge`), we use the vite environments API to fetch the corresponding transformed code for this module from the `ssr` environment.

In builds, we first build a bundle for the `ssr` environment, using the SSR bridge module as the entry point. Then we build the `worker` environment, and  have the `worker` environment consume the already built `ssr` environment (as if it were a library).

In dev, we make it so that any of the modules in import chains starting with the SSR bridge module are also loaded via the `ssr` environment. In other words, the `worker` environment gets a subgraph of ssr modules, each one corresponding to a module in the `ssr` environment. To accomplish this, we apply a wrapper around the code we fetch from the `ssr` environment for each modules and prepend a `virtual:rwsdk:ssr:` prefix to each import at runtime. We then tell vite how to load these virtual modules - by fetching them from the ssr environment:

```tsx
const transformedCode = `
await (async function(__vite_ssr_import__, __vite_ssr_dynamic_import__) {${code}})(
  (id) => __vite_ssr_import__('/@id/${VIRTUAL_SSR_PREFIX}'+id),
  (id) => __vite_ssr_dynamic_import__('/@id/${VIRTUAL_SSR_PREFIX}'+id));
`;
```

This way, for both dev and builds, SSR gets its own module resolution and transformation applied via the `ssr` environment, while the `worker` still gets to process the results to make it work within cloudflare worker environments.

## Supporting `react-server` import conditions
We can simply define different configuration for the different vite environments: the `worker` environment includes the `react-server` condition in its resolution configuration, while the `ssr` environment does not.

### Supporting `"use client` in `node_modules`
First we need to find client modules present in `node_modules`. We already do an initial scan for `"use client"`` directives. This was changed to include `node_modules` in this scan.

The next hurdle is `optimizeDeps`: In dev, vite runs dependencies through esbuild - this serves both as an optimization so that we can get bundles for the deps rather than needing to resolve chains of dependencies each time, but also to de-CJS-ify modules. We need to make sure the transformations we apply for `use client` (and `use server`) modules also work in the context of `optimizeDeps`.

For this reason, we apply the transformations both via the vite `transform()` hook (for user modules in dev, and all modules in builds), and via an esbuild plugin we configure for `optimizeDeps` (for deps during dev). We also need to add some `optimizeDeps` aliasing so that we can be sure the client and server component entry points we find in the scan are seen as the same modules esbuild finds when bundling the relevant deps themselves (so that we can end up with common chunk bundles, rather than accidental duplicate code for these modules).

## Notes for users
This is a backwards compatible change. However, this is also a significant change in approach for how RedwoodSDK bundles for RSC. It is very possible that there will be some instability at first while we get the approach working correctly for everyone.

For this reason, we'll first be doing preview releases of this change while we get to something stable, starting with `0.1.0-alpha`. We'll still be maintaining `0.0.x` releases until then - you're free to continue with `0.0.x` releases, or try out `0.1.0` preview releases. Hotfixes will reach both these preview releases and `0.0.x` releases.

## Issues solved
Fixes #429 
Fixes #449 
Fixes #448