fix: Separate entry point for client SSR (#494)
## Problem
Currently, if `rwsdk/client` is imported in a `"use client"` component, this will cause us to evaluate code paths that were intended to only run in the browser in the cloudflare worker runtime environment too.

This plays out in unexpected ways. For example, `globalThis.__webpack_require__` (used for loading client components on both client side and SSR in worker) gets overridden to use the module loader meant for the browser.

## Solution
* Remove the global `workerd` import condition from vite config: this way, resolutions in `client` environment now (correctly) do not follow `workerd` import conditions, but `ssr` and `worker` environments still do
* Change our `rwsdk/client` export conditions to point to a different entry point in the `workerd` case, so that SSR can get its own entry point for this specifier that avoids code paths we only want to run client side