feat: Warmup DO DB (#539)
Changes `createDb()` to start a warmup of the relevant durable object instance on first call. This allows us to run migrations as soon as possible for the DO, such that by the time DB interaction needs to happen, the DB is up and ready.

More context on what this DO db thing is all about coming soon!