fix: Ensure we resolve react relative to rwsdk package (#541)
## Background context
We currently have a vite plugin for ensuring react resolves to the sdk's react deps. We do this so that we have control over what version react rwsdk is on in order to control stability while RSC and rwsdk are both being more stable.

## Problem
We were doing `enforce: 'post'` for this vite plugin, and so some other plugin's resolveId was reaching resolutions for react before this plugin was. As a result, we'd end up getting resolving react to modules we weren't expecting, such as ones that might exist above the project dir if there are `node_modules` above the project dir.

## Solution
Use `enforce: 'pre'` so that we're first in the chain.

## Links
* User issue: https://discord.com/channels/679514959968993311/1382564500527517706/1382725155322724394