fix: Fix gaps in rwsdk/db type inference (#567)
We had an incomplete implementation of kysely's API's type definitions. This PR adds the missing methods - now via assertions so we can be sure we've covered everything.

It also fixes bugs in the type inference that were causing type inference for renames of columns and tables to break in some cases (e.g. several renames, or renames back to original name).