Add way to interrupt response handling for client side navigation. (#603)
When client-side navigation is initialized, we use fetch to request a new RFC payload. This RSC payload can also be an invalid response. 
In which case, client-side navigation breaks. This PR includes the ability to manually handle the response and determine whether it's an error, and then to just forward the browser directly to that page so that the user gets the proper experience. 

```tsx
const { handleResponse } = initClientNavigation()
initClient({ handleResponse })
```

**edit(justinvdm, 22 Jul 2025):**
This PR also now includes the changes needed to get this same functionality to work with the realtime: interruptors + client navigation + realtime client:

```tsx
const { handleResponse } = initClientNavigation()
initRealtimeClient({ handleResponse })
```

The main changes on the realtime side:
* refactoring
  * to avoid repetition on worker and client side: `ACTION_*` messages (when a client triggers an action and we're getting the corresponding response) and `RSC_*` messages (when server sends up an update - e.g. as a result of a different client's triggering an action) now reuse logic where possible
  * packing and unpacking protocol messages now encapsulated in `packMessage` + `unpackMessage` + types for protocol messages
* This refactoring allowed to make way for
  * including a status code to use to rehydrate `Response`s over websocket on client side - without refactoring this turned out to be surprisingly difficult
  * for both `ACTION_*` and `RSC_*` messages, allow for short-circuiting processing of responses using `handleResponses` - to integrate with client navigation

Fixes #601