feat: Allow response status and headers to be modified
I think that we need to consider adding this as an option. I don't know if this is the right way to go about it, but I wanted to start the conversation.

The `renderToString` and `renderToStream` methods are, I think, quite difficult to get right - and this is a quick win for us so that we can unblock @jldec.

I think I found better names for the request info object. I think we should call it HTTP messages. Which is what the Firefox manual refers to as request and response. I think we should also deprecate the headers and then just require that people set it on the actual response object.